Refactor products
